ログの型定義を整理
