+++ src/entrypoints/background/comment-filter/save-log.ts
@@ -1,27 +1,26 @@
 import type { FilteredData } from "./filter-comment";
 import { sumNumbers } from "@/utils/util";
-import type { CommentFiltering } from "@/types/storage/log-comment.types";
 import { colors } from "@/utils/config";
 import { mergeCount, setLog } from "@/utils/db";
 import type { ConditionalKeys } from "type-fest";
 import type { RuleFilter } from "./rule-filter";
 import { getRuleFilters } from "./rule-filter";
 import { setBadgeState } from "@/utils/browser";
-import type { Count } from "@/types/storage/log.types";
+import type { Count, LogData } from "@/types/storage/log.types";
 
 export async function saveLog(
     filteredData: FilteredData,
     logId: string,
     tabId: number,
 ) {
+    const comment = createCommentLog(filteredData);
     const count = createCount(filteredData);
-    const filtering = createFiltering(filteredData);
 
     await Promise.all([
         setLog(
             async () => {
                 return {
-                    commentFilterLog: { filtering },
+                    comment,
                     count: await mergeCount(count, logId),
                 };
             },
@@ -30,6 +29,25 @@ export async function saveLog(
         ),
         setBadgeState(count.blockedComment, colors.commentBadge, tabId),
     ]);
+}
+
+export function createCommentLog(
+    filteredData: FilteredData,
+): NonNullable<LogData["comment"]> {
+    const filteredComments = Object.values(filteredData.filters).flatMap(
+        (filter) => filter.getFilteredComments(),
+    );
+    const renderedComments = filteredData.threads.flatMap((thread) =>
+        thread.comments.map(({ body, userId, score }) => {
+            return { body, userId, score };
+        }),
+    );
+
+    return {
+        strictUserIds: filteredData.strictUserIds,
+        filteredComments,
+        renderedComments,
+    };
 }
 
 export function createCount(filteredData: FilteredData) {
@@ -53,21 +71,4 @@ export function createCount(filteredData: FilteredData) {
         exclude: calc("getExcludeCount"),
         disable: filters.commandFilter.getDisableCount(),
     } satisfies Count;
-}
-
-export function createFiltering(filteredData: FilteredData): CommentFiltering {
-    const filteredComments = Object.values(filteredData.filters).flatMap(
-        (filter) => filter.getFilteredComments(),
-    );
-    const renderedComments = filteredData.threads.flatMap((thread) =>
-        thread.comments.map(({ body, userId, score }) => {
-            return { body, userId, score };
-        }),
-    );
-
-    return {
-        strictUserIds: filteredData.strictUserIds,
-        filteredComments,
-        renderedComments,
-    };
 }
