+++ src/entrypoints/background/message.ts
@@ -209,7 +209,7 @@ async function getCommentsForDropdown(sender: browser.runtime.MessageSender) {
     const userId = dropdownComment?.userId;
     if (userId === undefined) return;
 
-    return log.commentFilterLog?.filtering?.renderedComments
+    return log.comment?.renderedComments
         .filter((comment) => comment.userId === userId)
         .toSorted((a, b) => a.body.localeCompare(b.body))
         .toSorted((a, b) => a.score - b.score)
